LayoutTests/media/modern-media-controls/tracks-support/audio-single-track-expected.txt

Check that no tracks button is shown when there's only one audio track.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('button.overflow') became different from null
PASS shadowRoot.querySelector('button.overflow').getBoundingClientRect().width became different from 0
Tapping overflow button...
PASS contextmenu.some((item) => item.title === 'Playback Speed') is true
Dismissing contextmenu...
PASS shadowRoot.querySelector('button.tracks') became different from null
PASS shadowRoot.querySelector('button.tracks').getBoundingClientRect().width became 0
PASS successfullyParsed is true

TEST COMPLETE
